refactor: move components from /drawing/inline to /drawing  - they will be used for other positioning element (floating)
