+++ src/file/drawing/doc-properties/doc-properties-attributes.ts
