+++ src/file/drawing/doc-properties/doc-properties.ts
