+++ src/file/drawing/effect-extent/effect-extent-attributes.ts
