+++ src/file/drawing/effect-extent/effect-extent.ts
