+++ src/file/drawing/graphic-frame/graphic-frame-locks/graphic-frame-locks.ts
