+++ src/file/drawing/graphic-frame/graphic-frame-properties.ts
