+++ src/file/drawing/graphic/graphic-data/graphic-data-attribute.ts
