+++ src/file/drawing/graphic/graphic-data/index.ts
