+++ src/file/drawing/graphic/graphic-data/pic/blip/blip-fill.ts
