+++ src/file/drawing/graphic/graphic-data/pic/blip/source-rectangle.ts
