+++ src/file/drawing/graphic/graphic-data/pic/blip/stretch.ts
