+++ src/file/drawing/graphic/graphic-data/pic/index.ts
