+++ src/file/drawing/graphic/graphic-data/pic/non-visual-pic-properties/non-visual-pic-properties.ts
