+++ src/file/drawing/graphic/graphic-data/pic/non-visual-pic-properties/non-visual-properties/non-visual-properties.ts
