+++ src/file/drawing/graphic/graphic-data/pic/pic-attributes.ts
