+++ src/file/drawing/graphic/graphic-data/pic/pic.ts
