+++ src/file/drawing/graphic/graphic-data/pic/shape-properties/no-fill.ts
