+++ src/file/drawing/graphic/graphic-data/pic/shape-properties/outline/no-fill.ts
