+++ src/file/drawing/graphic/graphic-data/pic/shape-properties/outline/outline.ts
