+++ src/file/drawing/graphic/graphic-data/pic/shape-properties/preset-geometry/adjustment-values/adjustment-values.ts
