+++ src/file/drawing/graphic/graphic-data/pic/shape-properties/preset-geometry/preset-geometry.ts
