+++ src/file/drawing/graphic/graphic-data/pic/shape-properties/shape-properties-attributes.ts
